ios-14.2.yaml: add yet another receipt sample
Also adapts the ReceiptExchange online test. Apparently, Apple stores
the number of keys created per device for a limited time only if running
in the development environment. Against that background, the test now
only verifies that the risk metric has a value of one or greater.

In the production environment, one should never observe a decreasing
risk metric value for the same device.