ReceiptValidator: rename `notAfter` to `notBefore`
So far, the functions `validateReceiptAsync` and `validateReceipt` had
been accepting an argument `notAfter` which denoted a timestamp. If a
receipt contains a timestamp which is before this deadline, it is
considered expired. Therefore, the argument should have the name
`notBefore`. This is also in line with X.509 certificates which use the
same name to mark the beginning of the validity period.

Please note that the actual logic is unchanged and has been working
correctly.